requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
geojson==3.2.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
oauthlib==3.2.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dotenv==1.0.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
unattended-upgrades==0.1
urllib3==2.3.0
uWSGI==2.0.28
virtualenv==20.29.1
virtualenv-clone==0.5.7
wadllib==1.3.6
Werkzeug==3.1.3
zipp==1.0.0
